requirements.txt
python-telegram-bot[job-queue]==21.9
python-dotenv==1.0.1
pytz==2025.1
